+++ src/main/java/lotto/controller/dto/LottoNumbersDto.java
@@ -1,4 +1,4 @@
-package lotto.application.dto;
+package lotto.controller.dto;
 
 import java.util.List;
 import lotto.domain.vo.Lotto;
