+++ src/main/java/lotto/view/ConsoleInputView.java
@@ -1,4 +1,4 @@
-package lotto.ui;
+package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
 import java.util.Arrays;
